refactor error_utils
