+++ Makefile
@@ -41,9 +41,7 @@ SRC			= $(SRC_DIR)/main.c \
 				$(SRC_DIR)/wrapper/xfork.c \
 				$(SRC_DIR)/wrapper/xunlink.c \
 				$(SRC_DIR)/utils/pipe_utils.c \
-				$(SRC_DIR)/utils/append_str.c \
-				$(SRC_DIR)/utils/error_exit.c \
-				$(SRC_DIR)/utils/usage_error_exit.c
+				$(SRC_DIR)/utils/error_utils.c
 OBJ			= $(patsubst $(SRC_DIR)/%.c, $(BUILD_DIR)/%.o, $(SRC))
 DEP			= $(patsubst $(SRC_DIR)/%.c, $(BUILD_DIR)/%.d, $(SRC))
 HEADER		= $(wildcard $(INC_DIR)/*.h)
