+++ src/utils/error_utils.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   usage_error_exit.c                                 :+:      :+:    :+:   */
+/*   error_utils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/30 14:45:07 by reasuke           #+#    #+#             */
-/*   Updated: 2024/05/31 14:51:09 by reasuke          ###   ########.fr       */
+/*   Created: 2024/06/13 19:26:33 by reasuke           #+#    #+#             */
+/*   Updated: 2024/06/13 19:48:34 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,39 @@
 
 #include "constants.h"
 #include "libft.h"
+#include "utils.h"
+
+static void	_append_str(char **dst, const char *src)
+{
+	ft_strlcat(*dst, src, ft_strlen(*dst) + ft_strlen(src) + 1);
+}
+
+static char	*_create_error_message(const char *func, const char *msg)
+{
+	char	*err_msg;
+	size_t	len;
+
+	len = ft_strlen(PROG_NAME) + ft_strlen(": ") + ft_strlen(func)
+		+ ft_strlen(": ") + ft_strlen(msg);
+	err_msg = ft_xmalloc(len + 1);
+	_append_str(&err_msg, PROG_NAME);
+	_append_str(&err_msg, ": ");
+	_append_str(&err_msg, func);
+	_append_str(&err_msg, ": ");
+	_append_str(&err_msg, msg);
+	_append_str(&err_msg, "\n");
+	return (err_msg);
+}
+
+void	error_exit(const char *func, const char *msg, int status)
+{
+	char	*err_msg;
+
+	err_msg = _create_error_message(func, msg);
+	write(STDERR_FILENO, err_msg, ft_strlen(err_msg));
+	free(err_msg);
+	exit(status);
+}
 
 void	usage_error_exit(void)
 {
